+++ packages/bpk-component-accordion/src/withAccordionItemState-test.js
@@ -2,35 +2,35 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import renderer from 'react-test-renderer';
 
-import withState from './withState';
 import BpkAccordionItem from './BpkAccordionItem';
+import withAccordionItemState from './withAccordionItemState';
 
-const StatefulBpkAccordionItem = withState(BpkAccordionItem);
+const EnhancedComponent = withAccordionItemState(BpkAccordionItem);
 
-describe('withState', () => {
+describe('withAccordionItemState(BpkAccordionItem)', () => {
   it('should render correctly', () => {
     const tree = renderer.create(
-      <StatefulBpkAccordionItem id="my-accordion" title="My accordion item">
+      <EnhancedComponent id="my-accordion" title="My accordion item">
         My accordion content
-      </StatefulBpkAccordionItem>,
+      </EnhancedComponent>,
     ).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('should render correctly with "expanded" prop', () => {
     const tree = renderer.create(
-      <StatefulBpkAccordionItem id="my-accordion" title="My accordion item" expanded>
+      <EnhancedComponent id="my-accordion" title="My accordion item" expanded>
         My accordion content
-      </StatefulBpkAccordionItem>,
+      </EnhancedComponent>,
     ).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('should render correctly with "initiallyExpanded" prop', () => {
     const tree = renderer.create(
-      <StatefulBpkAccordionItem id="my-accordion" title="My accordion item" initiallyExpanded>
+      <EnhancedComponent id="my-accordion" title="My accordion item" initiallyExpanded>
         My accordion content
-      </StatefulBpkAccordionItem>,
+      </EnhancedComponent>,
     ).toJSON();
     expect(tree).toMatchSnapshot();
   });
@@ -38,9 +38,9 @@ describe('withState', () => {
 
   it('should toggle "expanded" on click', () => {
     const accordionItem = shallow(
-      <StatefulBpkAccordionItem id="my-accordion" title="My accordion item" initiallyExpanded>
+      <EnhancedComponent id="my-accordion" title="My accordion item" initiallyExpanded>
         My accordion content
-      </StatefulBpkAccordionItem>,
+      </EnhancedComponent>,
     );
 
     expect(accordionItem.state('expanded')).toBe(true);